+++ src/test/java/roomescape/acceptance/admin/AuthAcceptanceTest.java
@@ -20,7 +20,7 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 import static roomescape.acceptance.Fixture.secretKey;
-import static roomescape.acceptance.PreInsertedData.ADMIN;
+import static roomescape.PreInsertedData.ADMIN;
 
 class AuthAcceptanceTest extends BaseAcceptanceTest {
