+++ src/test/java/roomescape/acceptance/admin/ReservationCreationAcceptanceTest.java
@@ -20,7 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static roomescape.acceptance.Fixture.adminToken;
-import static roomescape.acceptance.PreInsertedData.*;
+import static roomescape.PreInsertedData.*;
 
 @DisplayName("관리자가 예약을 추가한다.")
 class ReservationCreationAcceptanceTest extends BaseAcceptanceTest {
