+++ src/test/java/roomescape/acceptance/admin/ReservationDeletionAcceptanceTest.java
@@ -18,7 +18,7 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static roomescape.acceptance.Fixture.*;
-import static roomescape.acceptance.PreInsertedData.*;
+import static roomescape.PreInsertedData.*;
 
 class ReservationDeletionAcceptanceTest extends BaseAcceptanceTest {
