+++ src/test/java/roomescape/acceptance/admin/ThemeAcceptanceTest.java
@@ -19,8 +19,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static roomescape.acceptance.Fixture.adminToken;
-import static roomescape.acceptance.PreInsertedData.THEME_1;
-import static roomescape.acceptance.PreInsertedData.THEME_2;
+import static roomescape.PreInsertedData.THEME_1;
+import static roomescape.PreInsertedData.THEME_2;
 
 class ThemeAcceptanceTest extends BaseAcceptanceTest {
